MyChessStats/requirements.txt
Django>=5.1.5,<5.2
djangorestframework>=3.15.2,<3.16
python-chess>=1.999,<2.0
django-cors-headers>=4.3.1,<5.0
django-bootstrap5>=23.3,<24.0
Pillow>=10.2.0,<11.0